Fix various issues discovered via php-handlebars
* Improper rv usage in array_find, lookup_on_context, and resolve_possible_lambda
* Copy data ref in setup_options
* Compiler and VM should use same type for flags
* Add handlebars_compiler_get_known_helpers
* Fix issue with invoke_ambiguous
* Use handlebars_throw_ex in a couple places where possible
* Don't skip iterator init when user count is zero
* Immortalize strings in opcode serializer
* Remove warn unused result from stack pop

+++ src/handlebars_vm.h
@@ -89,7 +89,7 @@ struct handlebars_value * handlebars_vm_call_helper_str(
     struct handlebars_value * rv
 ) HBS_ATTR_NONNULL(1, 4, 5) HBS_ATTR_WARN_UNUSED_RESULT;
 
-void handlebars_vm_set_flags(struct handlebars_vm * vm, unsigned flags) HBS_ATTR_NONNULL_ALL;
+void handlebars_vm_set_flags(struct handlebars_vm * vm, unsigned long flags) HBS_ATTR_NONNULL_ALL;
 void handlebars_vm_set_helpers(struct handlebars_vm * vm, struct handlebars_value * helpers) HBS_ATTR_NONNULL_ALL;
 void handlebars_vm_set_partials(struct handlebars_vm * vm, struct handlebars_value * helpers) HBS_ATTR_NONNULL_ALL;
 void handlebars_vm_set_data(struct handlebars_vm * vm, struct handlebars_value * data) HBS_ATTR_NONNULL_ALL;
